# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TOAN Store E-commerce API
  description: |
    API documentation for TOAN Store E-commerce platform.
    
    **Authentication**: JWT via HTTP-Only Cookies (`nike_auth_session` / `nike_admin_session`)  
    **Rate Limiting**: Redis-backed sliding window, Fail-Closed for sensitive endpoints.
  version: 2.5.0
  contact:
    email: admin@toanstore.com
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/reviews:
    get:
      tags: [Reviews]
      summary: List reviews
      parameters:
        - { name: productId, in: query, schema: { type: integer } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: sort, in: query, schema: { type: string, enum: [newest, highest, lowest] } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Reviews]
      summary: Create review (must have purchased)
      security: [{ cookieAuth: [] }]
      responses:
        '201': { description: Review created }

  # ═══════════════════════════════════
  # PROMOTIONS
  # ═══════════════════════════════════
...
  /api/promo-codes/validate:
    post:
      tags: [Promotions]
      summary: Validate a promo code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, cartTotal]
              properties:
                code: { type: string }
                cartTotal: { type: number }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nike_auth_session
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: nike_admin_session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        data: {}
        message: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...